moves workflow testing code to separate go file
The import path is unaffected by this,
because it still is in the same go package.
